Rename async to coroutine to avoid keyword
